28 api add require user and admin

+++ api/orders.js
@@ -5,12 +5,16 @@ const {
   getOrdersByUser,
   getOrdersByStatus,
   createOrder,
+  setOrderAsPaymentPending,
+  setOrderAsProcessing,
+  setOrderAsSuccess,
   updateOrder,
-  deleteOrder
+  deleteOrder,
 } = require("../db");
-const { requireAdmin } = require("./utils");
+const { requireAdmin, requireUser } = require("./utils");
 
-ordersRouter.get("/", requireAdmin, async (req, res, next) => {
+// only admins should be allowed to see all exisitng orders
+ordersRouter.get("/", requireUser, requireAdmin, async (req, res, next) => {
   try {
     const orders = await getAllOrders();
     res.send(orders);
@@ -23,6 +27,7 @@ ordersRouter.get("/", requireAdmin, async (req, res, next) => {
   }
 });
 
+//any users should be able to look up their order by order id
 ordersRouter.get("/:orderId", async (req, res, next) => {
   const { orderId } = req.params;
   try {
@@ -37,51 +42,57 @@ ordersRouter.get("/:orderId", async (req, res, next) => {
   }
 });
 
-// TODO: require admin
-ordersRouter.get("/username/:username", requireAdmin, async (req, res, next) => {
-  try {
-    const { username } = req.params;
-    const orders = await getOrdersByUser(username);
-    console.log("orders by username: ", orders);
-    res.send(orders);
-  } catch (error) {
-    console.error(error);
-    next({
-      name: "NoExistingOrders",
-      message: "There are no orders under that username",
-    });
-  }
-});
-
-ordersRouter.get("/status/:status", requireAdmin, async (req, res, next) => {
-  try {
-    const { status } = req.params;
-    const orders = await getOrdersByStatus(status);
-    res.send(orders);
-  } catch (error) {
-    console.error(error);
-    next({
-      name: "orderDoesNotExist",
-      message: "There are no orders matching status",
-    });
-  }
-});
-
-// TODO: require user
-ordersRouter.post("/add", async (req, res, next) => {
-  try {
-    const { userId, email, address, status } = req.body;
-    if (!userId || !email || !address || !status) {
+//any registered users should be able to pull all of their orders
+ordersRouter.get("/username/:username", requireUser, async (req, res, next) => {
+    try {
+      const { username } = req.params;
+      const orders = await getOrdersByUser(username);
+      res.send(orders);
+    } catch (error) {
+      console.error(error);
       next({
-        name: "orderMissingFields",
-        message: "Please fill in the required field",
+        name: "NoExistingOrders",
+        message: "There are no orders under that username",
       });
-    } else {
+    }
+  }
+);
+
+// only admins should be allowed to see all exisitng orders by status
+ordersRouter.get("/status/:status", requireUser, requireAdmin, async (req, res, next) => {
+    const { status } = req.params;
+    try {
+      const orders = await getOrdersByStatus(status);
+      res.send(orders);
+    } catch (error) {
+      console.error(error);
+      next({
+        name: "orderDoesNotExist",
+        message: "There are no orders matching status",
+      });
+    }
+  }
+);
+
+// Create a new order on visit and when there is no existing cart. Create a new order when the order status changes to success.
+// Order default status is order_pending.
+// endpoint "/cart"
+ordersRouter.post("/", async (req, res, next) => {
+  const { email, address } = req.body;
+  try {
+    if (!req.user) {
       const newOrder = await createOrder({
-        userId,
+        userId: 1,
         email,
         address,
-        status,
+      });
+      res.send(newOrder);
+    } else if (req.user) {
+      const { id } = req.user;
+      const newOrder = await createOrder({
+        userId: id,
+        email,
+        address,
       });
       res.send(newOrder);
     }
@@ -94,21 +105,72 @@ ordersRouter.post("/add", async (req, res, next) => {
   }
 });
 
-// TODO: require user and checkOwner OR require admin and checkAdmin
+// endpoint "/checkout". Order status changes to "payment_pending". During this step, users should fill in their payment method such as billing address, credit card info, etc."
+ordersRouter.patch("/checkout", async (req, res, next) => {
+  try {
+    const { id } = req.body;
+    const orderStatus = await setOrderAsPaymentPending(id);
+    res.send(orderStatus);
+  } catch (error) {
+    console.error(error);
+    next({
+      name: "OrderStatusProcessingError",
+      message: "Failed to update the order as processing",
+    });
+  }
+});
+
+// endpoint "/pay" once user confirms pay. Order status changes to "processing."
+ordersRouter.patch("/pay", async (req, res, next) => {
+  try {
+    const { id } = req.body;
+    const orderStatus = await setOrderAsProcessing(id);
+    res.send(orderStatus);
+  } catch (error) {
+    console.error(error);
+    next({
+      name: "OrderStatusSuccessError",
+      message: "Failed to update the order as success",
+    });
+  }
+});
+
+// endpoint "/confirm". Order status changes to "success" once admin manually confirms the order's payment.
+ordersRouter.patch("/confirm", requireUser, requireAdmin, async (req, res, next) => {
+  try {
+    const { id } = req.body;
+    const orderStatus = await setOrderAsSuccess(id);
+    res.send(orderStatus);
+  } catch (error) {
+    console.error(error);
+    next({
+      name: "OrderStatusSuccessError",
+      message: "Failed to update the order as success",
+    });
+  }
+});
+
+// This is to update any order info such as email and address. This is unrelated to status updates.
 ordersRouter.patch("/update/:orderId", async (req, res, next) => {
   const { orderId } = req.params;
-  console.log("orderId", orderId);
-  const { email, address, currentStatus } = req.body;
+  const { id, isAdmin } = req.user;
   try {
-    const updatedOrder = await updateOrder({
-      id: orderId,
-      email,
-      address,
-      currentStatus,
-    });
-    console.log("updatedorder", updatedOrder);
-    res.send(updatedOrder);
-    return;
+    const { userId } = await getOrderById(orderId);
+    if (id === userId || isAdmin) {
+      const { email, address } = req.body;
+      const updatedOrder = await updateOrder({
+        id: orderId,
+        userId: userId,
+        email,
+        address,
+      });
+      res.send(updatedOrder);
+    } else {
+      next({
+        name: "InvalidUserError",
+        message: "You are not the owner of this account",
+      });
+    }
   } catch (error) {
     console.error(error);
     next({
@@ -118,35 +180,21 @@ ordersRouter.patch("/update/:orderId", async (req, res, next) => {
   }
 });
 
-// to update the status only (to process orders)
-ordersRouter.patch("/status/:orderId", async (req, res, next) => {
+ordersRouter.delete("/:orderId", requireUser, async (req, res, next) => {
   const { orderId } = req.params;
-  console.log("orderId", orderId);
-  const { currentStatus } = req.body;
+  const { id, isAdmin } = req.user;
   try {
-    const updatedOrder = await updateOrder({
-      id: orderId,
-      currentStatus,
-    });
-    res.send(updatedOrder);
-    return;
-  } catch (error) {
-    console.error(error);
-    next({
-      name: "StatusUpdateError",
-      message: "Failed to update the order",
-    });
-  }
-});
-
-// TODO: require user and checkOwner OR require admin and checkAdmin
-ordersRouter.delete("/:orderId", async (req, res, next) => {
-  const { orderId } = req.params;
-  console.log("orderId", orderId);
-  try {
-    const updatedOrder = await deleteOrder(orderId);
-    res.send(updatedOrder);
-    return;
+    const { userId } = await getOrderById(orderId);
+    if (id === userId || isAdmin) {
+      const updatedOrder = await deleteOrder(orderId);
+      res.send(updatedOrder);
+    } else {
+      next({
+        name: "InvalidUserError",
+        message:
+          "You are not the owner of this account or do not have the permission to update the status",
+      });
+    }
   } catch (error) {
     console.error(error);
     next({
